修改我的页面和设置页面

+++ videoutils_lib/src/main/java/com/lansosdk/videoeditor/LanSongFileUtil.java
@@ -409,6 +409,7 @@ public class LanSongFileUtil {
                 e.printStackTrace();
             }
         } else {
+            if (dir.listFiles() == null) return;
             for (File file : dir.listFiles()) {
                 if (file.isFile()) {
                     file.delete();
